See if this fixes vm_addr
We guess at the IP address of the host by querying
`virsh domifaddr --full <name>`. However, sometimes this gives multiple
addresses (such as if the guest updated its /etc/machine-id in the case
of a systemd-networkd/systemd-resolved host like Ubuntu). Grab the
last one in the hopes its the right one.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>